Add Employee method AddressMap to call view with Google Maps Embeded view of selected address
AddressMap method on employee controller creates a URL that will be called by javascript within the AddressMap view.
Url is stored in Employee NotMapped multi-purpose Dictionary<string, string> Address to store the string of the address and
Google Maps Embeded API is used as a src for a iframe, which is currently set to a bootstrap embeded-fluid class.
Can used to show multiple maps if needed.